+++ modules/app-control/src/color.js
@@ -7,7 +7,7 @@ export function rgb(color) {
   if (cache[color]) return cache[color];
   const parsed = HSV.exec(color)[1]
     .split(",")
-    .map(i => parseInt(i));
+    .map(i => parseInt(i, 10));
   const c = hsvToRgb.apply(null, parsed);
   cache[color] = "rgb(" + c[0] + "," + c[1] + "," + c[2] + ")";
   return cache[color];