+++ modules/app-control/src/load-set.hoc.jsx
@@ -1,34 +1,28 @@
 import React from "react";
+import { loadAudioSet, getEvents } from "@atpls/audioset";
+import initKeyboard from "@atpls/keyboard";
 import Message from "./components/Message";
-import keyboard from "@atpls/keyboard";
-import Events from "nanobus";
 
 export default function loadSetHoc(WrappedComponent) {
   class LoadSetComponent extends React.Component {
     constructor(props) {
       super(props);
-      this.socket = props.socket;
       this.state = { set: null, events: null };
     }
     componentWillMount() {
       const { url, socket } = this.props;
-      fetch(url + "sets/continentes.audioset.json")
-        .then(response => response.json())
-        .then(set => {
-          const events = initSet(set, socket);
-          this.setState({ set, events });
-        });
+      const setName = window.location.hash.slice(1).toLowerCase() || "lik03";
+      loadAudioSet(setName, url + "sets/").then(set => {
+        const events = getEvents(set, socket);
+        initKeyboard(set, events);
+        this.setState({ set, events });
+      });
     }
 
     render() {
       const { set, events } = this.state;
       return set ? (
-        <WrappedComponent
-          set={set}
-          events={events}
-          socket={this.socket}
-          {...this.props}
-        />
+        <WrappedComponent set={set} events={events} {...this.props} />
       ) : (
         <Message label="Loading antropoloops set..." />
       );
@@ -36,10 +30,4 @@ export default function loadSetHoc(WrappedComponent) {
   }
 
   return LoadSetComponent;
-}
-
-function initSet(set, socket) {
-  const events = new Events();
-  keyboard(set, events);
-  return events;
 }