+++ modules/app-control/src/ws-connect.hoc.jsx
@@ -1,24 +1,7 @@
 import React from "react";
-import io from "socket.io-client";
+import { initConnection, serverUrl } from "@atpls/audioset";
 import Message from "./components/Message";
 import Connect from "./components/Connect";
-
-// try to get a valid server url
-const serverUrl = () => {
-  const origin = window.origin || "http://127.0.0.1:3333";
-  const portIndex = origin.lastIndexOf(":");
-  return (portIndex === -1 ? origin : origin.slice(0, portIndex)) + ":3333";
-};
-
-const loadStatus = url =>
-  fetch(url)
-    .then(response => response.json())
-    .then(status => {
-      console.log("Status loaded!", status);
-      const url = status.address;
-      const socket = io(url);
-      return { url, socket };
-    });
 
 /**
  * A HighOrderComponent to connect via web sockets
@@ -30,8 +13,7 @@ export default function wsConnectHoc(WrappedComponent) {
       this.state = { error: null, socket: null, url: null };
     }
     componentWillMount() {
-      const url = serverUrl() + "/status";
-      loadStatus(url)
+      initConnection(window.origin)
         .then(this.setState.bind(this))
         .catch(error => {
           this.setState({ error });
@@ -43,9 +25,7 @@ export default function wsConnectHoc(WrappedComponent) {
           <Connect
             onConnect={baseUrl => {
               this.setState({ error: undefined });
-              const url = baseUrl + "/status";
-              console.log("loading", url);
-              loadStatus(url).then(state => this.setState(state));
+              initConnection(baseUrl).then(state => this.setState(state));
             }}
           />
         );