+++ modules/webapp/src/audio-set/index.js
@@ -1,6 +1,0 @@
-// import Set from "./set";
-import Manager from "./manager";
-
-export function AudioSetManager() {
-  return new Manager();
-}